requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
alembic==1.8.0
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.2
blinker==1.4
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fakeredis==1.8.1
fastapi==0.78.0
fastapi-mail==1.0.9
fastapi-mailman==0.0.4
fastapi-pagination==0.9.3
Flask==2.1.2
Flask-Mail==0.9.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.9.1
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
redis==4.3.3
requests==2.28.0
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.37
starlette==0.19.1
stripe==3.5.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
